+++ src/scripts/lobby/components/atoms/SelectAvatarTable/SelectAvatarTableBodyAvatar/SelectAvatarTableBodyAvatarImage.tsx
@@ -8,7 +8,7 @@ interface Props {
   readonly navigatable: boolean
 }
 
-export const SelectAvatarTableBodyImage: React.FC<Props> = props => {
+export const SelectAvatarTableBodyAvatarImage: React.FC<Props> = props => {
   const handleSelect = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
     if (event.target === event.currentTarget) {
       props.handleSelect()
@@ -29,4 +29,4 @@ export const SelectAvatarTableBodyImage: React.FC<Props> = props => {
     </div>
   )
 }
-SelectAvatarTableBodyImage.displayName = 'SelectAvatarTableBodyImage'
+SelectAvatarTableBodyAvatarImage.displayName = 'SelectAvatarTableBodyAvatarImage'
