+++ src/scripts/lobby/components/atoms/SelectAvatarTable/SelectAvatarTableBodyAvatar/SelectAvatarTableBodyAvatarName.test.tsx
