+++ src/scripts/lobby/components/atoms/SelectAvatarTable/SelectAvatarTableBodyAvatar/SelectAvatarTableBodyAvatarName.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import {TextInput} from '../Input/TextInput'
+import {TextInput} from '../../Input/TextInput'
 
 interface Props {
   readonly additionalClassName: string[]
